requirements.txt
asgiref==3.7.2 
django==4.2.3 
djangorestframework==3.14.0 
pytz==2023.3 
sqlparse==0.4.4 
typing-extensions==4.7.1 
tzdata==2023.3
python-dotenv==1.0.0
certifi==2023.5.7 
cffi==1.15.1 
charset-normalizer==3.2.0 
cryptography==41.0.2 
defusedxml==0.7.1 
django-templated-mail==1.1.1 
djangorestframework-simplejwt==5.2.2
djoser==2.2.0 
idna==3.4 
oauthlib==3.2.2 
pycparser==2.21 
pyjwt==2.7.0 
python3-openid==3.2.0 
requests==2.31.0 
requests-oauthlib==1.3.1 
social-auth-app-django==5.2.0 
social-auth-core==4.4.2 
urllib3==2.0.3
django_filter==23.2
psycopg2-binary==2.9.6
flake8==6.0.0 
mccabe==0.7.0 
pycodestyle==2.10.0 
pyflakes==3.0.1
